requirements.txt
numpy==1.24.4
packaging==23.1
pycryptodome==3.20.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
QDarkStyle==3.1
QtPy==2.4.0
six==1.16.0
tabulate==0.9.0
transliterate==1.10.2
